requirements.txt
mysql-connector-python
pillow
Flask
Flask-SQLAlchemy